refactor: move entities into server modules
