+++ apps/servers/src/auth/auth.controller.ts
@@ -3,7 +3,7 @@ import { JwtService } from '@nestjs/jwt';
 import * as argon2 from 'argon2';
 import { Repository } from 'typeorm';
 import { InjectRepository } from '@nestjs/typeorm';
-import { User } from '../../../libs/shared/types/user.entity';
+import { User } from '../users/user.entity';
 
 @Controller('auth')
 export class AuthController {
