--- libs/shared/types/index.ts
@@ -1,2 +1,0 @@
-export * from './user.entity';
-export * from './blood-test.entity';
