+++ tsconfig.base.json
@@ -5,6 +5,9 @@
     "lib": ["es2017"],
     "rootDir": ".",
     "strict": true,
+    "experimentalDecorators": true,
+    "emitDecoratorMetadata": true,
+    "strictPropertyInitialization": false,
     "esModuleInterop": true,
     "skipLibCheck": true,
     "forceConsistentCasingInFileNames": true
