+++ tsconfig.json
@@ -1,0 +1,8 @@
+{
+  "extends": "./tsconfig.base.json",
+  "compilerOptions": {
+    "outDir": "dist"
+  },
+  "include": ["apps/**/*", "libs/**/*"],
+  "exclude": ["node_modules"]
+}
